sign commit

+++ CodeON/src/main/java/com/spring/app/entity/Department.java
@@ -21,7 +21,7 @@ public class Department {
         initialValue = 10
     )
     @Column(name = "department_seq", nullable = false)
-    private int departmentSeq;
+    private long departmentSeq;
 
     @Column(name = "department_name", nullable = false, length = 30)
     private String departmentName;
